pyproject.toml
[project]
name = "intel-summarizer"
version = "0.1.0"
description = "Personal automated intelligence summarizer"
requires-python = ">=3.11"

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.26",
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "lxml_html_clean>=0.4",
    "aiohttp>=3.9",
    "httpx>=0.27",
    "anthropic>=0.39",
    "model2vec>=0.4",
    "scipy>=1.12",
    "python-telegram-bot>=21.0",
    "fpdf2>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["intel*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
